pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "flask-mysqldb>=2.0.0",
    "python-dotenv>=1.1.1",
]
